Rename react to components
