+++ src/components/highscore.js
