+++ src/index.js
@@ -1,16 +1,16 @@
 var Client = require('react-engine/lib/client');
 var Routes = require('./routes');
 
-require('./react/index');
-require('./react/layout');
-require('./react/room');
-require('./react/404');
+require('./components/index');
+require('./components/layout');
+require('./components/room');
+require('./components/404');
 
 var options = {
   routes: Routes,
   viewResolver: function (viewName) {
     try {
-      return require('./react/' + viewName);
+      return require('./components/' + viewName);
     } catch (e) {
       console.log('The component ' + viewName + ' is not loaded in index.js');
     }
